+++ src/genie/core/parameter/descriptor_subseq_cfg.cc
@@ -116,7 +116,7 @@ bool DescriptorSubseqCfg::isClassSpecific() const { return class_specific_dec_cf
 // ---------------------------------------------------------------------------------------------------------------------
 
 void DescriptorSubseqCfg::write(util::BitWriter& writer) const {
-    writer.writeBits(static_cast<uint8_t>(class_specific_dec_cfg_flag), 1);
+    writer.writeBits(class_specific_dec_cfg_flag, 1);
     for (auto& i : descriptor_configurations) {
         i->write(writer);
     }
