+++ src/genie/quality/paramqv1/qv_coding_config_1.cc
@@ -49,7 +49,7 @@ void QualityValues1::setQvps(ParameterSet&& _parameter_set_qvps) {
 // ---------------------------------------------------------------------------------------------------------------------
 
 void QualityValues1::write(util::BitWriter& writer) const {
-    writer.writeBits(static_cast<uint8_t>(qv_coding_mode), 4);
+    writer.writeBits(qv_coding_mode, 4);
     writer.writeBits(static_cast<bool>(parameter_set_qvps), 1);
     if (parameter_set_qvps) {
         parameter_set_qvps->write(writer);
